login and register now working! updated controllers and public js login and withAuth
