+++ controllers/homePageRoutes.js
@@ -13,5 +13,18 @@
 
 const router = require('express').Router();
 
+// router.get('/', (req, res) => {
+   
+// });
+
+router.get('/login', (req, res) => {
+    // If the user is already logged in, redirect the request to another route
+    if (req.session.logged_in) {
+      res.redirect('/pets');
+      return;
+    }
+  
+    res.render('signUp');
+  });
 
 module.exports = router;
