+++ utils/withAuth.js
@@ -1,7 +1,7 @@
 const withAuth = (req, res, next) => {
     // If the user is not logged in, redirect the request to the login route
     if (!req.session.logged_in) {
-      res.redirect('/signup');
+      res.redirect('/login');
     } else {
       next();
     }
