+++ views/layouts/main.handlebars
@@ -62,9 +62,19 @@
           <a class="button is-primary">
             <strong>Sign up</strong>
           </a>
+
+  {{#if loggedIn}}
+          <a class="button is-light">
+            Log out
+          </a>
+
+  {{else}}
+
           <a class="button is-light">
             Log in
           </a>
+  {{/if}}
+
         </div>
       </div>
     </div>
@@ -74,10 +84,10 @@
       {{{body}}}
     </div>
   </main>
-  {{#if loggedIn}}
+  {{!-- {{#if loggedIn}} --}}
   {{!-- script for homepage/main page functionality --}}
   {{!-- <script src="/js/logout.js"></script> --}}
-  {{/if}}
+  {{!-- {{/if}} --}}
 </body>
 
 </html>
