+++ views/signUp.handlebars
@@ -39,4 +39,4 @@
   </div>
 </div>
 
-<script src="./js/login.js"></script>
+<script src="./js/login-signup.js"></script>
